+++ app/src/main/java/com/serge/moneyme/service/PMTHelper.kt
@@ -5,7 +5,7 @@ import kotlin.math.pow
 
 class PMTHelper {
     companion object {
-        fun calculatePMT(R: Double, n: Double, Pv: Double): Double {
+        fun calculatePMT(R: Double, n: Int, Pv: Double): Double {
             val pvTimesR = (Pv * R)
             val onePlusR = 1 + R
             val negativeN = n * -1
