+++ app/src/main/java/com/serge/moneyme/service/Persistence.kt
@@ -12,9 +12,9 @@ const val NAME_KEY = "nameKey"
 class Persistence(context: Context) {
     private val preferences: SharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
 
-    fun saveFinanceDetails(amount: Int, months: Int) {
-        preferences.edit().putInt(AMOUNT_KEY, amount).apply()
-        preferences.edit().putInt(MONTHS_KEY, months).apply()
+    fun saveFinanceDetails(amount: Float, months: Float) {
+        preferences.edit().putFloat(AMOUNT_KEY, amount).apply()
+        preferences.edit().putFloat(MONTHS_KEY, months).apply()
     }
 
     fun saveUserEmail(email: String) {
@@ -29,8 +29,8 @@ class Persistence(context: Context) {
         preferences.edit().clear().apply()
     }
 
-    fun getAmountValue() = preferences.getInt(AMOUNT_KEY, 0)
-    fun getMonthsValue() = preferences.getInt(MONTHS_KEY, 0)
+    fun getAmountValue() = preferences.getFloat(AMOUNT_KEY, 0.0F)
+    fun getMonthsValue() = preferences.getFloat(MONTHS_KEY, 0.0F)
     fun getEmailValue() = preferences.getString(EMAIL_KEY, "")
     fun getNameValue() = preferences.getString(NAME_KEY, "")
 }
